requirements.txt
tensorflow>=2.10.0
gradio>=4.0.0
opencv-python-headless>=4.8.0
numpy>=1.23.0
Pillow>=9.0.0
kaggle
tf-keras
